projects/project5_2/requirements.txt
crewai
crewai-tools
langchain_openai
python-dotenv
pyyaml
redis
grpcio
grpcio-tools
protobuf

